+++ db/migrate/20131021173445_create_cindices.rb
@@ -1,6 +1,6 @@
 class CreateCindices < ActiveRecord::Migration
   def change
-    drop_table :companies_indices
+    drop_table :companies_indices  if self.table_exists?("companies_indices")
     create_table :cindices do |t|
       t.references :company
       t.references :index
